requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
numpy==1.26.0
WeasyPrint==60.1
latex2mathml==3.77.0
python-dotenv==1.0.0
xhtml2pdf==0.2.16
